+++ src/main/java/org/mmarini/rltd/TDRelu.java
@@ -59,7 +59,7 @@ public class TDRelu extends TDLayer {
     }
 
     @Override
-    public INDArray[] train(INDArray[] inputs, INDArray output, INDArray grad, INDArray delta, TDNetwork net) {
+    public INDArray[] train(INDArray[] inputs, INDArray output, INDArray grad, INDArray delta, float lambda) {
         return new INDArray[]{
                 grad.mul(inputs[0].gt(0))
         };