+++ src/test/java/org/mmarini/rltd/TDSoftmaxTest.java
@@ -137,7 +137,7 @@ class TDSoftmaxTest {
         INDArray[] in = new INDArray[]{inputs};
         TDSoftmax layer = new TDSoftmax("name", temperature);
         INDArray out = layer.forward(in, null);
-        INDArray[] post_grads = layer.train(in, out, grad, Nd4j.zeros(1), null);
+        INDArray[] post_grads = layer.train(in, out, grad, Nd4j.zeros(1), 0);
 
         assertThat(post_grads, arrayWithSize(1));
         assertThat(post_grads[0], matrixCloseTo(new float[][]{{